Note because it is my intention to remove multiple data support in favor of DSTs
in the future. Therefore I have left some issues caused by this merge
unresolved, so some of the tests relating to multiple data types don't pass and
should be expected.

Conflicts:
	src/compiler-rt
	src/compiletest/compiletest.rs
	src/liblibc/lib.rs
	src/libnative/io/c_unix.rs
	src/libnative/io/file_unix.rs
	src/libnative/io/process.rs
	src/librustc/driver/config.rs
	src/librustc/lib/llvm.rs
	src/librustc/middle/dataflow.rs
	src/librustc/middle/trans/adt.rs
	src/librustc/middle/trans/base.rs
	src/librustc/middle/trans/cabi.rs
	src/librustc/middle/trans/debuginfo.rs
	src/librustc/middle/trans/expr.rs
	src/librustc/middle/trans/type_.rs
	src/librustc/middle/ty.rs
	src/librustc/middle/typeck/coherence.rs
	src/librustc/middle/typeck/infer/to_str.rs
	src/librustrt/unwind.rs
	src/libstd/lib.rs
	src/libstd/os.rs
	src/libstd/rt/backtrace.rs
	src/libsyntax/abi.rs
	src/llvm

+++ src/libcore/ops.rs
@@ -570,10 +570,11 @@ pub trait Shl<RHS,Result> {
 
 macro_rules! shl_impl(
     ($($t:ty)*) => ($(
-        #[cfg(not(test))]
         impl Shl<$t, $t> for $t {
             #[inline]
-            fn shl(&self, other: &$t) -> $t { (*self) << (*other) }
+            fn shl(&self, other: &$t) -> $t {
+                (*self) << (*other as uint)
+            }
         }
     )*)
 )
@@ -612,10 +613,9 @@ pub trait Shr<RHS,Result> {
 
 macro_rules! shr_impl(
     ($($t:ty)*) => ($(
-        #[cfg(not(test))]
         impl Shr<$t, $t> for $t {
             #[inline]
-            fn shr(&self, other: &$t) -> $t { (*self) >> (*other) }
+            fn shr(&self, other: &$t) -> $t { (*self) >> (*other as uint) }
         }
     )*)
 )
